+++ sdk-example/src/main/java/com/sportradar/livedata/sdk/example/LiveScoutFeedListenerImpl.java
@@ -79,10 +79,10 @@ public class LiveScoutFeedListenerImpl implements LiveScoutFeedListener {
         }
     }
 
-//    @Override
-//    public void onPartialMatchUpdateReceived(LiveScoutFeed sender, MatchUpdateEntity matchUpdate) {
-//        logger.info("Partial match update");
-//    }
+    @Override
+    public void onFullPaginatedMatchUpdateReceived(LiveScoutFeed sender, MatchUpdateEntity matchUpdate) {
+        logger.info("Full paginated match update");
+    }
 
     @Override
     public void onLineupsReceived(LiveScoutFeed sender, LineupsEntity lineups) {